% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWKT.R
\name{convertWKT}
\alias{convertWKT}
\title{convertWKT}
\usage{
convertWKT(wkt, epsg_from, epsg_to)
}
\arguments{
\item{wkt}{String: Valid representation of a WKT geometry}

\item{epsg_from}{String: The current EPSG number of the WKT e.g. 2193 for NZTM}

\item{epsg_to}{String: The EPGS number you want to convert to e.g. 4326 for WGS84}
}
\description{
Converts a WKT string from one EPSG to another
}
\examples{
new_wkt <- convertWKT(wkt = old_wkt, epsg_from = "2193", epsg_to = "4326")
}
