% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFilter.R
\name{createFilter}
\alias{createFilter}
\title{createFilter Function}
\usage{
createFilter(filter_field, filter_operator, filter_value)
}
\arguments{
\item{filter_field}{A valid name of a field in the table you are going to filter}

\item{filter_operator}{Any, EqualTo, NotEqualTo, GreaterThan, GreaterEqualThan, LessThan, LessEqualThan, In, NotIn, IsNull, IsNotNull, Like, NotLike, Matches}

\item{filter_value}{A value to include in the filter}
}
\description{
Creates a filter object that can be passed into the getTableData function
}
\examples{
# TODO:  Add example
}
