% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaItem.R
\name{downloadMultimedia}
\alias{downloadMultimedia}
\title{downloadMedia Function}
\usage{
downloadMultimedia(headers, table_name, asset_id, path)
}
\arguments{
\item{headers}{Authorised access headers created using  getHeaders(...).}

\item{table_name}{The name of the table the asset is in (must be a valid table name)}

\item{asset_id}{The id of the asset in the table you want to download}

\item{path}{The path to where all files will be downloaded to}
}
\description{
Downloads and saves all multimedia objects associated with a particular asset
You need to get the headers using the getHeaders() function
}
\examples{
\dontrun{
#The following example downloads all multimedia objects associated with sign id 1

sign_table <- getTableData(
headers=hdrs,
table_name="sign",
asset_id=1,
path="./temp_folder/"
)

}
}
