% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthorisedHeaders.R
\name{getAuthorisedHeaders}
\alias{getAuthorisedHeaders}
\title{getAuthorisedHeaders Function}
\usage{
getAuthorisedHeaders(userName, password, database)
}
\arguments{
\item{userName}{RAMM user login}

\item{password}{Password for your login}

\item{database}{Name of the RAMM database you want to connect to}
}
\description{
This function creates the headers for your requests to the RAMM API
}
\examples{
headers <- getAuthorisedHeaders(userName='api_demo', password='thursday', database='RAMM API Demo')
print(headers)
}
