% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClosestRoads.R
\name{getClosestRoads}
\alias{getClosestRoads}
\title{getClosestRoadsFunction}
\usage{
getClosestRoads(headers, wkt, search_distance = 50)
}
\arguments{
\item{headers}{A headers object created using the getHeaders function}

\item{wkt}{A valid WKT string in EPSG:4326}

\item{search_distance}{Distance in metres for search radius e.g. 50 = 50m search radius}
}
\description{
Retrieves a dataframe of the closest road ids and distances to a given WKT
Note if the WKT is a Polygon, it will give all roads that touch that WKT and will return a distance of 0.
The distance parameter is ignore if the WKT is a Polygon
Have not tested with a linestring as yet.
WKT must be in EPSG 4326.  Use the convertWKT function if required
}
\examples{
# The following example will find all roads within 1km of the point Lat: -45.32112 and Lng: 170.40021 on the database authorised through the headers object
closest_roads <- getClosestRoads(headers=headers, wkt="POINT(170.40021 -45.32112)", 1000)
}
