% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabaseDetails.R
\name{getDatabaseDetails}
\alias{getDatabaseDetails}
\title{getDatabaseDetails Function}
\usage{
getDatabaseDetails(headers)
}
\arguments{
\item{headers}{A headers object created using the getHeaders function}
}
\description{
This returns the details of the database you are connected to via the headers (which contains the token)
The returned object contains information regarding the map projection you may need to convert to/from
}
\examples{
dbDetails <- getDatabaseDetails(headers=getHeaders(...))
}
