% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeaders.R
\name{getHeaders}
\alias{getHeaders}
\title{getHeaders Function}
\usage{
getHeaders(userName, password, database)
}
\arguments{
\item{userName}{RAMM user login}

\item{password}{Password for your login}

\item{database}{Name of the RAMM database you want to connect to}
}
\description{
This function creates the headers for your requests to the RAMM API
}
\examples{
headers <- getHeaders(userName='xxxx', password='xxxxx', database='A RAMM database')
}
