% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableWithCustomTypes.R
\name{getTableWithCustomTypes}
\alias{getTableWithCustomTypes}
\title{getTableWithCustomTypes Function}
\usage{
getTableWithCustomTypes(headers, table_name, custom_types_df,
  load_type = "CoreAndLocation", named_columns = "", get_geometry = FALSE,
  filters = list())
}
\arguments{
\item{headers}{Authorised access headers created using  getHeaders(...).}

\item{table_name}{The name of the table you want to extract data from (must be a valid table name)}

\item{custom_types_df}{A dataframe of column names and the types that will be forced. needs to be named column_name and column_type}

\item{load_type}{Which columns of data will be returned - can be
'All', 'CoreAndLocation' (i.e. core columns with primary key) or 'Specified'.}

\item{named_columns}{If loadType = 'Specified' a character vector of required columns (must be valid column names).}

\item{get_geometry}{(Boolean) Return geometry with the data. Returns an SF class data.frame in NZTM coordiantes (i.e. EPSG/SRID 2193).}

\item{filters}{A \strong{LIST} of filters created using createFilter(...).}
}
\description{
Retrieve RAMM table and force column types - the schema type listing from the RAMM API is not always correct!
You need to get the headers using the getHeaders() function
}
