% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhsWidening.R
\name{offsetLeftHandSurfaces}
\alias{offsetLeftHandSurfaces}
\title{offsetLeftHandSurfaces Function}
\usage{
offsetLeftHandSurfaces(headers, site_road_id, site_start_m, site_end_m,
  site_offset)
}
\arguments{
\item{headers}{Authorised access headers created using  getHeaders(...).}

\item{site_road_id}{The RAMM Road ID on which the site resides}

\item{site_start_m}{The start chainage(m) of the site}

\item{site_end_m}{The end chainage(m) of the site}

\item{site_offset}{The width(m) of the offset that you want to implement at the site}
}
\description{
This function returns raw c_surface records from RAMM that need to be updated or added when doing a Left Hand Side widening
}
\examples{
\dontrun{
#The following example returns dataframes of the adds and updates when widening road_id 888 between 120m and 400m by 4m

surfaces = offsetLeftHandSurfaces(headers = hdrs, site_road_id = 888, start_start_m = 120, site_end_m = 400, site_offset = 4)
surfaces$adds
surfaces$updates
}
}
