% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableData.R
\name{getTableData}
\alias{getTableData}
\title{getTableData Function}
\usage{
getTableData(headers, table_name, named_columns, get_geometry = FALSE,
  expand_lookups = FALSE, filters = list())
}
\arguments{
\item{headers}{A headers object created using the getHeaders function}

\item{table_name}{The name of the table you want to extract data from (must be a valid table name)}

\item{named_columns}{A vector of all of the columns you want to return (must be valid column names)}

\item{get_geometry}{Return a geometry(TRUE) or not (FALSE).  Will append the field "wkt_geom" to your returned data}

\item{expand_lookups}{Automatically return the inbuilt lookup names instead of the parent value e.g. road_id returns road name}

\item{filters}{A list of filters created using the createFilter function}
}
\description{
Retrieves data from the specified RAMM table from the authorised database (as per the token in the headers).
You need to get the headers using the getHeaders() function
}
\examples{
#The following example returns the raw road_id, sign_type, and wkt_geom fields from the sign table of the database authorised through the headers

sign_table <- getTableData(headers=headers, table_name="sign",named_columns=c("road_id","sign_type"),get_geometry=TRUE, expand_lookups=FALSE)
}
